feat(app): accept a fractional Glk host cell size, rounded per window (SQ-1603)
TerminalFacts::glk_cell_px / AppState::glk_cell_px / AppGlk::char_px are
now (f64, f64) instead of (u32, u32), so a host with a genuinely
fractional text cell (e.g. 10.2x22.95 device pixels) no longer has to
round before calling into lanthorn at all. canvas_size rounds each open
graphics window's own `cells x char_px` independently, at the
multiplication, rather than sharing one pre-rounded ratio across every
window — closing the cumulative drift a single rounded ratio produces
across windows of different cell counts.

GlkBackend::char_pixels() (the gvm trait boundary a game's own bytecode
observes via glk_window_get_size) stays integer and rounds once, since
gvm's own layout math is unaffected and out of scope here — only AppGlk's
own storage and canvas_size's multiply gained precision.

Threading char_px as f64 through GlulxSession's constructors touched
every call site across crates/app/src and crates/app/tests that passed a
literal cell-size tuple; each was converted mechanically (parsed by call
site, not by blind regex) and re-verified against cargo check.

New discriminating test (sq1598_glk_cell_px.rs) proves per-window
rounding against real Kerkerkruip graphics windows of different cell
counts at char_px = (10.2, 22.95), where global-ratio rounding and
per-window rounding disagree by a different amount on each window
(11px/9px/3px) — falsified by temporarily routing canvas_size through a
single pre-rounded ratio and confirming the new test fails.

Quest: SQ-1603
Completes: SQ-1603

+++ crates/app/src/host/screen.rs
@@ -90,7 +90,11 @@ pub fn resize_glulx(session: &mut dyn Engine, (cols, rows): (u16, u16)) -> bool 
 /// graphics canvas to `(window cells) × char_px` and drives the game's own
 /// redraw the same way [`resize_glulx`] does for a terminal-size change —
 /// there is no separate mechanism. `false` for any other engine.
-pub fn set_glk_cell_px(session: &mut dyn Engine, char_px: (u32, u32)) -> bool {
+///
+/// Fractional, matching `TerminalFacts::glk_cell_px` (SQ-1603) — each open
+/// canvas rounds its own `cells × char_px` independently rather than sharing
+/// one pre-rounded ratio; see `AppGlk::canvas_size`.
+pub fn set_glk_cell_px(session: &mut dyn Engine, char_px: (f64, f64)) -> bool {
     match session.as_any_mut().downcast_mut::<GlulxSession>() {
         Some(gs) => {
             gs.set_char_px(char_px);
